Failed attempt to implement Strassen matrix multiplication on recursive layout as described in Siddhartha Chatterjee, Alvin R. Lebeck, Praveen K. Patnala and Mithuna Thottehodi paper "Recursive Array Layout and Fast Matrix Multiplication".
As of 2009-02-13, this implementation does not work! The padding at left and bottom
sides of the matrix should be cleared after some operations like scalerAdd and is not.
Also there is a limitation in the multiplication that can only process matrices with
sizes similar enough to have the same power of two number of tiles in all three
matrices A, B and C such that C = A*B. These parts have not been fixed since the
performance gain with respect to DenseRealMatrix are not very important, and the
numerical stability is poor.

This may well be due to a bad implementation.

This code has been put in the experimental directory for the record, putting it into
production would require solving all these issues.


git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@744126 13f79535-47bb-0310-9956-ffa450edef68