Added checkstyle filter for exceptional catch of RuntimeException.
These catch occur in multi-start optimizers. The exceptions are caught
and stored, and if all restarts fail, then one of the stored
RuntimeException is rethrown. The purpose of multi-start is both to
allow some trials to fail and to avoid being trapped in a local
extremum. Catching RuntimeException is therefore mandatory here, and not
dangerous since we rethrow them as needed.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1179947 13f79535-47bb-0310-9956-ffa450edef68