Added lower and upper boundaries to finite differences.
When a function is defined only on an interval, finite differences
should not attempt to compute sample points outside this interval. This
case is now detected properly and the sample is shifted if needed. This
may result in some loss of accuracy as the formula is not centered
anymore, but at least we should not get catastrophic errors or
exceptions.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1416249 13f79535-47bb-0310-9956-ffa450edef68