Added a new ExtendedFieldElement interface.
This interface represents anything that is real number like. It is
implemented by Decimal64, Dfp and DerivativeStructure. The purpose of
this interface is to be able to set up general algorithms that apply on
real numbers (maybe complex too) and use genericity.

A first use case corresponds to 3D geometry objects (Vector3D and
Rotation).

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1449529 13f79535-47bb-0310-9956-ffa450edef68